// apps/web/src/generated/api/models/freeboardCreateRequest.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * SoSo API 명세서
 * 소소한 아이디어 공유 플랫폼의 백엔드 API 문서입니다.
 * OpenAPI spec version: v1.0.0
 */
import type { FreeboardCreateRequestCategory } from './freeboardCreateRequestCategory';

/**
 * 자유게시판 글 작성 요청
 */
export interface FreeboardCreateRequest {
  /** 게시글 카테고리 */
  category: FreeboardCreateRequestCategory;
  /**
   * 게시글 제목
   * @minLength 0
   * @maxLength 100
   */
  title: string;
  /**
   * 게시글 내용
   * @minLength 0
   * @maxLength 5000
   */
  content: string;
  /**
   * 첨부 이미지 파일들 (최대 4장)
   * @minItems 0
   * @maxItems 4
   */
  images?: Blob[];
}