// apps/web/src/generated/api/models/freeboardDetailResponseCategory.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * SoSo API 명세서
 * 소소한 아이디어 공유 플랫폼의 백엔드 API 문서입니다.
 * OpenAPI spec version: v1.0.0
 */

/**
 * 카테고리
 */
export type FreeboardDetailResponseCategory =
  (typeof FreeboardDetailResponseCategory)[keyof typeof FreeboardDetailResponseCategory];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FreeboardDetailResponseCategory = {
  'daily-hobby': 'daily-hobby',
  restaurant: 'restaurant',
  'living-convenience': 'living-convenience',
  'neighborhood-news': 'neighborhood-news',
  startup: 'startup',
  others: 'others',
} as const;